refactor: 스크롤 동작 커스텀 훅 생성
